backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
cryptography==42.0.0
opencv-python==4.9.0.80
deepface==0.0.79
easyocr==1.7.0
Pillow<10.0.0
python-multipart==0.0.6
rasa==3.6.0
alembic==1.13.0
python-dotenv==1.0.0
tensorflow==2.15.0
numpy==1.24.3
requests==2.31.0
pydantic==2.5.3
httpx==0.26.0
